Separate Model and View
Remove the logic from inside the Map class and place it inside its own unique class. This new class is responsible for going through the list of codes in the database and extracting their location information, then adding each point as a marker to an overlay.

This overlay is returned to the caller, and can later be added overtop the displayed map.
